Refactor xoffsetview into xfunctorview and functor parameter
